moved authentification stuff into the index.js
